requirements.txt
django
pillow
djangorestframework
django-filter
psycopg2
# djmail
# django-jinja
python-decouple
djangorestframework-simplejwt #JWT token


#pip3 install -r requirements.txt (run command to install these libraries)